# openapi.yaml
openapi: 3.0.3
info:
  title: Ruderalis API
  description: Medical marijuana ecommerce backend API
  version: 1.0.0
  contact:
    name: Ruderalis API Support
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
security: []
components:
  securitySchemes:
    UserAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated users (vendor or user role)
    VendorAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated vendors (role=vendor)
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated admins (admin or super-admin role)
  schemas:
...
    Price:
      type: object
      properties:
        amount:
          type: number
        currency:
          type: string
          enum: [GHS, USD, EUR]
    Stock:
      type: object
      properties:
        available:
          type: boolean
        quantity:
          type: integer
    SaleInfo:
      type: object
      properties:
        status:
          type: boolean
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        price:
          $ref: '#/components/schemas/Price'
    FeaturedInfo:
      type: object
      properties:
        status:
          type: boolean
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
    RatingDetails:
      type: object
      properties:
        five:
          type: integer
        four:
          type: integer
        three:
          type: integer
        two:
          type: integer
        one:
          type: integer
    Rating:
      type: object
      properties:
        average:
          type: number
        count:
          type: integer
        details:
          $ref: '#/components/schemas/RatingDetails'
    ProductMetadata:
      type: object
      properties:
        marijuana:
          type: object
        edible:
          type: object
        accessory:
          type: object
    Product:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        owner:
          type: string
          description: ObjectId reference to User
        shop:
          type: string
          description: ObjectId reference to Shop
        status:
          type: string
          enum: [active, deleted, pending]
        stock:
          $ref: '#/components/schemas/Stock'
        price:
          $ref: '#/components/schemas/Price'
        description:
          type: string
        image:
          type: string
        sale:
          $ref: '#/components/schemas/SaleInfo'
        rank:
          type: integer
        featured:
          $ref: '#/components/schemas/FeaturedInfo'
        metadata:
          $ref: '#/components/schemas/ProductMetadata'
        rating:
          $ref: '#/components/schemas/Rating'
        variant:
          type: string
          enum: [marijuana, edible, accessory]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Review:
      type: object
      properties:
        _id:
          type: string
        user:
          type: string
          description: ObjectId reference to User
        product:
          type: string
          description: ObjectId reference to Product
        text:
          type: string
        rating:
          type: number
          minimum: 0
          maximum: 5
        visible:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        default: 20
...
paths:
...
  /user/products:
    post:
      tags: [User Products]
      summary: Create a product
      security:
        - UserAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, shop, stock, price, description, image, details, variant]
              properties:
                name:
                  type: string
                shop:
                  type: string
                stock:
                  $ref: '#/components/schemas/Stock'
                price:
                  $ref: '#/components/schemas/Price'
                description:
                  type: string
                image:
                  type: string
                details:
                  type: object
                variant:
                  type: string
                  enum: [marijuana, edible, accessory]
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
    get:
      tags: [User Products]
      summary: List products
      security:
        - UserAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: owner
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Products retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  count:
                    type: integer
...
  /user/reviews:
    post:
      tags: [User Reviews]
      summary: Create a product review
      security:
        - UserAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product, text, rating]
              properties:
                product:
                  type: string
                text:
                  type: string
                rating:
                  type: number
                  minimum: 0
                  maximum: 5
      responses:
        '201':
          description: Review created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
    get:
      tags: [User Reviews]
      summary: List product reviews
      security:
        - UserAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: product
          in: query
          schema:
            type: string
        - name: visible
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Reviews retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
                  count:
                    type: integer
...